# governance/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/experiments/{run_id}/questions/{question_id}/feedback": {
      "put": {
        "tags": [
          "feedback"
        ],
        "summary": "Upsert Feedback",
        "description": "Create-or-merge feedback for one question. Returns the merged doc.",
        "operationId": "upsert_feedback_experiments__run_id__questions__question_id__feedback_put",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upsert Feedback Experiments  Run Id  Questions  Question Id  Feedback Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "feedback"
        ],
        "summary": "Delete Feedback",
        "operationId": "delete_feedback_experiments__run_id__questions__question_id__feedback_delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{run_id}/feedback": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Get Run Feedback",
        "description": "All feedback for one run, keyed by question_id — one call powers both\nthe run page and the copy-to-clipboard button (no N+1 per question).",
        "operationId": "get_run_feedback_experiments__run_id__feedback_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Run Feedback Experiments  Run Id  Feedback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{run_id}/questions/{question_id}/feedback/triage": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Triage Feedback",
        "description": "Feedback triage: classifies the CURRENT stored\ncomment for this question into structured error classes via the\nRealm's own local generator, deterministically cross-checks the\nextraction against this run's own retrieval trace (no LLM involved in\nthat half), and proposes a minimal-radius lever from the taxonomy.\n\nRead-only recommendation only —\nfirst stage: this endpoint never creates a pin or changes\nanything about the run itself. It records a PROPOSAL on the feedback\ndocument; a human confirms, edits, or rejects it via the `/decide`\nendpoint below, and creating an actual pin from a confirmed proposal is\na separate, explicit call to `POST /pins`.",
        "operationId": "triage_feedback_experiments__run_id__questions__question_id__feedback_triage_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Triage Feedback Experiments  Run Id  Questions  Question Id  Feedback Triage Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/{run_id}/questions/{question_id}/feedback/triage/decide": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Decide Triage",
        "description": "Records a human's decision on an existing triage proposal — never\ncreates a pin or any other side effect by itself (see\n`triage_feedback`'s own docstring for why).",
        "operationId": "decide_triage_experiments__run_id__questions__question_id__feedback_triage_decide_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "realm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Realm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageDecideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Decide Triage Experiments  Run Id  Questions  Question Id  Feedback Triage Decide Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackWriteRequest": {
        "properties": {
          "rating": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "good",
                  "bad"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Scores",
            "default": {}
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          }
        },
        "type": "object",
        "title": "FeedbackWriteRequest",
        "description": "All fields optional — a PUT only touches the fields it sets. `scores`\nmerges by key (setting `accuracy` alone must not clear an already-saved\n`completeness`) rather than replacing the whole dict. No fixed set of\nscore dimensions is enforced here — the UI renders `accuracy`/\n`completeness`/`relevance` by default, but this stays an open dict so a\nnew dimension can be added later without a schema migration (same\nphilosophy as `question_type` in datasets.py: a free string, not an\nenum)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TriageDecideRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "confirm",
              "reject",
              "edit"
            ],
            "title": "Action"
          },
          "edited_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Result"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "TriageDecideRequest",
        "description": "Body for `POST .../feedback/triage/decide` — a human's verdict on a\ntriage proposal (the improvement plan's first stage: the agent\n1\": the agent proposes, a human decides). `edited_result` is required\nonly for `action=\"edit\"` — a partial dict merged over the stored\ntriage_result (e.g. correcting `error_classes` the model got wrong)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}